Current control and position control are now implemented. The motor driver can now track trajectories. The project is complete, but the code needs to cleaned up.

+++ encoder.h
@@ -3,7 +3,7 @@
 
 void initEncoder();
 int getEncoderCount();
-int getEncoderDegrees();
+float getEncoderDegrees();
 void resetEncoder();
 
 #endif
